Rename dismissKeyboard file to handleKeyboard. Put viewMoving function in it
